Improve QML part
- Add toolbar buttons
- Refactor project structure
