+++ CMakeLists.txt
@@ -3,12 +3,13 @@
 
 cmake_minimum_required(VERSION 3.22)
 
-project(AnimBoom VERSION 0.1 LANGUAGES C CXX)
+project(AnimBoomDesigner VERSION 0.1 LANGUAGES C CXX)
 
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
+set(CMAKE_AUTOMOC ON)
 include(${CMAKE_CURRENT_LIST_DIR}/cmake/Options.cmake)
 
 # TODO: Change it to target_compile_options
@@ -37,38 +38,47 @@ endif()
 
 include(${CMAKE_CURRENT_LIST_DIR}/cmake/Macros.cmake)
 
-qt_standard_project_setup(REQUIRES 6.5)
+qt6_standard_project_setup(REQUIRES 6.5)
 
-add_subdirectory(modules)
-add_subdirectory(ui)
-
-qt_add_executable(AnimBoomDesigner
-    main.cpp
+qt6_add_executable(${CMAKE_PROJECT_NAME}
+    src/main.cpp
 )
 
-target_link_libraries(AnimBoomDesigner
+qt6_add_resources(${CMAKE_PROJECT_NAME} "configuration"
+    PREFIX "/"
+    FILES
+        qtquickcontrols2.conf
+)
+
+qt6_add_qml_module(${CMAKE_PROJECT_NAME}
+    URI "Main"
+    VERSION 1.0
+    NO_PLUGIN
+    RESOURCE_PREFIX
+        /
+    QML_FILES
+        main.qml
+)
+
+add_subdirectory(ui)
+add_subdirectory(src)
+
+target_link_libraries(${CMAKE_PROJECT_NAME}
     PRIVATE
+        Qt6::Core
+        Qt6::Gui
+        Qt6::Qml
         Qt6::Quick
         uiplugin
         animboom-toolkit
 )
 
-target_include_directories(AnimBoomDesigner
+target_include_directories(${CMAKE_PROJECT_NAME}
     PRIVATE
     ${CMAKE_CURRENT_BINARY_DIR}/include
 )
 
-add_subdirectory(assets)
-
-set_target_properties(AnimBoomDesigner PROPERTIES
-    MACOSX_BUNDLE_GUI_IDENTIFIER animboom.com
-    MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
-    MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
-    MACOSX_BUNDLE TRUE
-    WIN32_EXECUTABLE TRUE
-)
-
-install(TARGETS AnimBoomDesigner
+install(TARGETS ${CMAKE_PROJECT_NAME}
     BUNDLE DESTINATION .
     LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
