--- assets/CMakeLists.txt
@@ -1,9 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-only
-# Copyright (C) 2023 Krzysztof Tyburski
-
-qt_add_resources(AnimBoomDesigner "assets"
-    PREFIX "/assets"
-    FILES
-        logo/logo.ico
-        logo/logo.png
-)
