+++ cmake/FindPackages.cmake
@@ -2,5 +2,5 @@
 # Copyright (C) 2023 Krzysztof Tyburski
 
 # Qt packages
-find_package(Qt6 6.5 REQUIRED COMPONENTS Quick)
+find_package(Qt6 6.5 REQUIRED COMPONENTS Core Gui Qml Quick)
