+++ main.qml
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-only
+// Copyright (C) 2023 Krzysztof Tyburski
+
+import QtQuick
+import ui
+
+App {
+}
