+++ qtquickcontrols2.conf
