+++ ui/CMakeLists.txt
@@ -1,21 +1,27 @@
 # SPDX-License-Identifier: GPL-3.0-only
 # Copyright (C) 2023 Krzysztof Tyburski
 
-qt_add_library(ui STATIC)
-qt_add_qml_module(ui
-    URI ui
+qt6_add_library(ui STATIC)
+qt6_add_qml_module(ui
+    URI "ui"
     VERSION 1.0
-    QML_FILES
-        Main.qml
-        MainView.qml
-        MainViewForm.ui.qml
     RESOURCE_PREFIX
         /
+    QML_FILES
+        App.qml
+        MainView.qml
+        MainViewForm.ui.qml
+    RESOURCES
+        assets/logo/logo.ico
+        assets/logo/logo.png
+        assets/toolbar/ToolbarBackground.png
 )
 
+add_subdirectory(Utils)
+add_subdirectory(Widgets)
 
-qt_add_resources(ui "config"
-    PREFIX "/"
-    FILES
-        qtquickcontrols2.conf
+
+target_link_libraries(ui PRIVATE
+    Widgetsplugin
+    Utilsplugin
 )
