+++ ui/Utils/Constants.qml
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-only
+// Copyright (C) 2023 Krzysztof Tyburski
+
+pragma Singleton  
+import QtQuick  
+  
+QtObject {
+    readonly property int width: 1920
+    readonly property int height: 1080
+}
