+++ ui/Utils/FontContainer.qml
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: GPL-3.0-only
+// Copyright (C) 2023 Krzysztof Tyburski
+
+pragma Singleton  
+import QtQuick  
+  
+QtObject {
+    readonly property FontLoader libreFranklinMedium: FontLoader { source: "fonts/LibreFranklin/LibreFranklin-Medium.ttf" }
+}
