+++ ui/Widgets/ToolbarButton.ui.qml
@@ -1,0 +1,84 @@
+
+
+/*
+This is a UI file (.ui.qml) that is intended to be edited in Qt Design Studio only.
+It is supposed to be strictly declarative and only uses a subset of QML. If you edit
+this file manually, you might introduce QML code that is not supported by Qt Design Studio.
+Check out https://doc.qt.io/qtcreator/creator-quick-ui-forms.html for details on .ui.qml files.
+*/
+
+// SPDX-License-Identifier: GPL-3.0-only
+// Copyright (C) 2023 Krzysztof Tyburski
+import QtQuick
+import QtQuick.Controls
+import Utils
+
+Button {
+    id: control
+
+    implicitWidth: Math.max(
+                       buttonBackground ? buttonBackground.implicitWidth : 0,
+                       textItem.implicitWidth + leftPadding + rightPadding)
+    implicitHeight: Math.max(
+                        buttonBackground ? buttonBackground.implicitHeight : 0,
+                        textItem.implicitHeight + topPadding + bottomPadding)
+    leftPadding: 4
+    rightPadding: 4
+
+    text: "Button"
+
+    width: 60
+    height: 30
+
+    background: buttonBackground
+    Rectangle {
+        id: buttonBackground
+        color: "#00000000"
+        implicitWidth: 100
+        implicitHeight: 40
+        opacity: enabled ? 1 : 0.3
+        radius: 20
+    }
+
+    contentItem: textItem
+    Text {
+        id: textItem
+        text: control.text
+
+        font: FontContainer.libreFranklinMedium.font
+        opacity: enabled ? 1.0 : 0.3
+        color: "#ffffff"
+        horizontalAlignment: Text.AlignHCenter
+        verticalAlignment: Text.AlignVCenter
+    }
+
+    states: [
+        State {
+            name: "normal"
+            when: !control.down
+
+            PropertyChanges {
+                target: buttonBackground
+                color: "#00000000"
+            }
+
+            PropertyChanges {
+                target: textItem
+                color: "#ffffff"
+            }
+        },
+        State {
+            name: "down"
+            when: control.down
+            PropertyChanges {
+                target: textItem
+                color: "#ffffff"
+            }
+
+            PropertyChanges {
+                target: buttonBackground
+                color: "#383838"
+            }
+        }
+    ]
+}
