+++ ui/Widgets/qmldir
@@ -1,0 +1,2 @@
+Module widgets
+ToolbarButton 1.0 ToolbarButton.ui.qml
